added uniform buffers bug

+++ src/Renderer/vcr_swap_chain.cpp
@@ -42,7 +42,7 @@ SwapChain::~SwapChain() {
         swapChain = nullptr;
     }
 
-    for (int i = 0; i < depthImages.size(); i++) {
+    for (size_t i = 0; i < depthImages.size(); i++) {
         vkDestroyImageView(device.device(), depthImageViews[i], nullptr);
         vkDestroyImage(device.device(), depthImages[i], nullptr);
         vkFreeMemory(device.device(), depthImageMemorys[i], nullptr);
@@ -304,7 +304,7 @@ void SwapChain::createDepthResources() {
     depthImageMemorys.resize(imageCount());
     depthImageViews.resize(imageCount());
 
-    for (int i = 0; i < depthImages.size(); i++) {
+    for (size_t i = 0; i < depthImages.size(); i++) {
         VkImageCreateInfo imageInfo{};
         imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
         imageInfo.imageType = VK_IMAGE_TYPE_2D;
